css:청첩장

+++ pages/component/MyContext.js
@@ -5,6 +5,7 @@ export const DbContext = createContext(null);
 const MyContext = ({children}) => {
 
     const [data,setData] = useState();
+    const [todoData,todoSetData] = useState();
 
     async function dataFun(type,obj){
         let trans;
@@ -18,19 +19,40 @@ const MyContext = ({children}) => {
                 console.log(res);
             })
             return dataFun('get');
-        }else{
+        }else if (type === 'delete'){
             await axios.delete(`/api/${obj}`)
             return dataFun('get');
         }
-        //console.log(trans);
         setData(trans);
     }
+
+    async function todoDataFun(type,obj){
+        let trans;
+        if(type=='get'){
+            await axios.get('/api/todo').then(res=>trans = res.data);
+        }else if(type=='post'){
+            await axios.post('/api/todo',obj)
+            return todoDataFun('get');
+        }else if(type=='put'){
+            await axios.put(`/api/todo/${obj.id}`, obj).then((res)=>{
+                console.log(res);
+            })
+            return dataFun('get');
+        }else if (type === 'delete'){
+            await axios.delete(`/api/todo/${obj}`)
+            return todoDataFun('get');
+        }
+        console.log(trans)
+        todoSetData(trans);
+    }
+
     useEffect(()=>{
-        dataFun('get')
+        dataFun('get');
+        todoDataFun("get");
     },[])
 
     return (
-        <DbContext.Provider value={{data,dataFun}}>
+        <DbContext.Provider value={{data,dataFun,todoData,todoDataFun}}>
             {children}
         </DbContext.Provider>
     )
